Adds a graph builder
